Successfully render smoke sequences, keyframed, improved!
Time dilation maxes out at 20, so I set that time dilation for the
niagara system using a construction/event blueprint. To get the
render working properly, I had to then set the sequencer time
dilation to 1, and render every 20th frame.

To get the sequencer to respect changes in position of the smoke,
I think I had to add a transform track... not sure. It was weird.